charlock and anagrams
Given a string , find the number of "unordered anagrammatic pairs" of
substrings. In other words, find the number of unordered pairs of
substrings of  that are anagrams of each other.

Two strings are anagrams of each other if the letters of one string can
be rearranged to form the other string.